feat: simplify login
